Move client-side Java code to `client` module and server-side utils under `server.util` package.
Moved some utility methods between classes.
