+++ client/src/test/java/org/spine3/protobuf/MessageFieldShould.java
