+++ examples/src/main/java/org/spine3/examples/aggregate/server/EventLogger.java
@@ -28,7 +28,7 @@ import org.spine3.examples.aggregate.event.OrderCreated;
 import org.spine3.examples.aggregate.event.OrderLineAdded;
 import org.spine3.examples.aggregate.event.OrderPaid;
 import org.spine3.protobuf.Messages;
-import org.spine3.util.Identifiers;
+import org.spine3.server.util.Identifiers;
 
 /**
  * A sample of event handler class that logs some of the events.
