+++ examples/src/main/java/org/spine3/examples/eventstore/EventPublisher.java
@@ -30,8 +30,8 @@ import org.spine3.base.EventContext;
 import org.spine3.base.EventId;
 import org.spine3.base.EventRecord;
 import org.spine3.server.stream.grpc.EventStoreGrpc;
-import org.spine3.util.EventRecords;
-import org.spine3.util.Events;
+import org.spine3.server.util.EventRecords;
+import org.spine3.server.util.Events;
 
 import java.util.concurrent.TimeUnit;
