+++ examples/src/main/java/org/spine3/examples/eventstore/EventReader.java
@@ -30,8 +30,8 @@ import org.spine3.base.EventContext;
 import org.spine3.base.EventRecord;
 import org.spine3.server.stream.EventStreamQuery;
 import org.spine3.server.stream.grpc.EventStoreGrpc;
+import org.spine3.server.util.EventRecords;
 import org.spine3.type.TypeName;
-import org.spine3.util.EventRecords;
 
 import java.util.Iterator;
 import java.util.concurrent.TimeUnit;
