+++ server/src/main/java/org/spine3/server/EntityId.java
@@ -25,7 +25,7 @@ import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableSet;
 import com.google.protobuf.Descriptors;
 import com.google.protobuf.Message;
-import org.spine3.util.Identifiers;
+import org.spine3.server.util.Identifiers;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.collect.Iterables.transform;
