+++ server/src/main/java/org/spine3/server/storage/StorageUtil.java
@@ -26,8 +26,8 @@ import com.google.protobuf.Any;
 import org.spine3.base.EventContext;
 import org.spine3.base.EventId;
 import org.spine3.base.EventRecord;
+import org.spine3.server.util.Events;
 import org.spine3.type.TypeName;
-import org.spine3.util.Events;
 
 import javax.annotation.Nullable;
 import java.util.List;
