+++ server/src/main/java/org/spine3/server/util/Identifiers.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.util;
+package org.spine3.server.util;
 
 import com.google.common.base.Function;
 import com.google.common.collect.ImmutableMap;
