+++ server/src/main/java/org/spine3/server/util/Methods.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.util;
+package org.spine3.server.util;
 
 import com.google.common.base.Predicate;
 import com.google.common.collect.ImmutableMap;
