+++ server/src/main/java/org/spine3/server/util/package-info.java
@@ -17,24 +17,12 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.spine3.util;
-
-import org.junit.Test;
-import org.spine3.base.EventId;
-
-import static org.junit.Assert.assertFalse;
 
 /**
- * @author Mikhail Melnik
+ * This package contains server-side utility classes.
  */
-@SuppressWarnings("InstanceMethodNamingConvention")
-public class EventsShould {
 
-    @Test
-    public void generate_event_id() {
-        final EventId result = Events.generateId();
+@ParametersAreNonnullByDefault
+package org.spine3.server.util;
 
-        assertFalse(result.getUuid().isEmpty());
-    }
-
-}
+import javax.annotation.ParametersAreNonnullByDefault;
