+++ server/src/test/java/org/spine3/server/util/MessagesShould.java
@@ -17,7 +17,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.spine3.util;
+package org.spine3.server.util;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.Message;
